fix: resolve bug where tests fail on codepen.io
Solution :remove useEffect and manually add handleClick to buttons
instead of using buttons.forEach

Explanation:
useEffect was needed to use buttons.forEach in
the component's body before the return statement which has the jsx for
the buttons. useEffect fires after the component has been rendered
and the jsx is returned.

However, adding clickEvents through buttons.forEach does not operate
the same in all environments for React and it's best avoided.